feat: 문자열 계산기 클래스 main으로 이동
