+++ src/test/kotlin/calculator/StringCalculatorTest.kt
@@ -1,5 +1,6 @@
 package calculator
 
+import calculate.StringCalculator
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.BeforeEach
@@ -8,37 +9,6 @@ import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 import org.junit.jupiter.params.provider.NullAndEmptySource
 import org.junit.jupiter.params.provider.ValueSource
-import java.util.regex.Pattern
-
-class StringCalculator {
-    fun inputText(text: String?): Int {
-        if (text.isNullOrBlank()) {
-            return 0
-        }
-
-        val matchText = CUSTOM_DELIMITER.matcher(text)
-        if (matchText.find()) {
-            return matchText.group(2).split(matchText.group(1))
-                .map { convertToInt(it) }
-                .sumOf { it }
-        }
-
-        return text.split(DEFAULT_DELIMITER)
-            .map { convertToInt(it) }
-            .sumOf { it }
-    }
-
-    private fun convertToInt(numberText: String): Int {
-        require(numberText.toInt() >= 0) { ERROR_MESSAGE }
-        return numberText.toInt()
-    }
-
-    companion object {
-        private val DEFAULT_DELIMITER: Pattern = Pattern.compile(",|:")
-        private val CUSTOM_DELIMITER: Pattern = Pattern.compile("//(.)\n(.*)")
-        private const val ERROR_MESSAGE: String = "음수는 입력할 수 없습니다."
-    }
-}
 
 class StringCalculatorTest {
     private lateinit var calculator: StringCalculator
